Renamed widget to UnreadItemDecoration to reflect how it's being used for now
